feat(habitat): move transformation transactions into domain
Move active pattern apply contracts, refusals, renderers, transaction input resolution, and worktree observation from src/lib/pattern-apply into src/domains/transformation-transaction. Update fix and transaction service modules to consume the domain directly and add architecture guards that keep the old lib feature path deleted.

OpenSpec: deep-habitat-effect-transformation-transaction-domain
