+++ openspec/changes/deep-habitat-effect-transformation-transaction-domain/design.md
@@ -2,24 +2,26 @@
 
 ## Owner
 
-Fix service module, with lower-level pattern apply, transformation transaction,
-and protected-zone internals still owning the current apply behavior.
+Fix service module and transformation transaction domain, with protected-zone
+internals still owning the current path authority behavior.
 
 `fix` is the command surface for Habitat-owned transformations. This slice
 moves command-level fix orchestration into that owned service module. The
-follow-on domain drain moves write-set, protected-zone, rollback, cleanup, and
-resource ownership into named domains.
+domain drain moves pattern apply contracts and transaction records out of
+generic `lib` ownership. Protected-zone, rollback, cleanup, and deeper
+write-resource ownership remain separate follow-on drains.
 
 ## Target Flow
 
 ```text
-Fix CLI -> Habitat service client -> fix service module -> pattern apply domain
+Fix CLI -> Habitat service client -> fix service module -> transformation transaction domain
 ```
 
 ## Write Set
 
 ```text
 tools/habitat-harness/src/service/modules/fix/**
+tools/habitat-harness/src/domains/transformation-transaction/**
 tools/habitat-harness/src/service/contract.ts
 tools/habitat-harness/src/service/router.ts
 tools/habitat-harness/src/commands/fix.ts
@@ -33,6 +35,7 @@ Deleted path:
 
 ```text
 tools/habitat-harness/src/lib/fix.ts
+tools/habitat-harness/src/lib/pattern-apply/**
 ```
 
 ## Required Cutover In This Slice
@@ -45,8 +48,13 @@ tools/habitat-harness/src/lib/fix.ts
 - Refuse `D0-package-export-symbol-runfix`; this slice does not add a
   replacement package helper export or a compatibility wrapper.
 - Route the CLI through `client.fix.run`.
+- Move pattern apply contracts, refusal records, renderers, transaction input
+  resolution, and worktree observation into
+  `src/domains/transformation-transaction/**`.
 - Keep dry-run and live-write intent projection stable.
 - Add tests that fail if the CLI imports or calls a `lib/fix` path directly.
+  Add tests that fail if active transaction logic remains under
+  `src/lib/pattern-apply/**`.
 
 ## Follow-On Transformation Domain Drain
 
@@ -54,7 +62,8 @@ tools/habitat-harness/src/lib/fix.ts
   protected zones, rollback data, and final disposition.
 - Grit dry-run output is not treated as write proof without local verification.
 - Temp/cache/write-set resources are scoped through Effect.
-- Pattern apply and protected-zone source move into named domains.
+- Pattern apply source moves into the transformation transaction domain.
+- Protected-zone source moves in a later protected-zone domain drain.
 
 ## Stop Conditions
