+++ openspec/changes/deep-habitat-effect-transformation-transaction-domain/specs/habitat-harness/spec.md
@@ -1,9 +1,10 @@
 ## ADDED Requirements
 
-### Requirement: Fix Runs Through Habitat Service Before Transaction Drain
+### Requirement: Fix Runs Through Habitat Service And Transaction Domain
 
-Habitat SHALL expose fix execution as an owned Habitat service module before
-the deeper transformation transaction and protected-zone authority drain.
+Habitat SHALL expose fix execution as an owned Habitat service module and SHALL
+own transformation transaction contracts in a named domain rather than generic
+`lib` feature logic.
 
 #### Scenario: Fix CLI runs
 
@@ -13,12 +14,15 @@ the deeper transformation transaction and protected-zone authority drain.
   selection, worktree observation, pattern transaction requests, and command
   stream output
 - **AND** package helper export `runFix` is not preserved as a wrapper
+- **AND** active pattern transaction contracts and renderers live under
+  `src/domains/transformation-transaction/**`, not `src/lib/pattern-apply/**`
 - **AND** the CLI handles only flags, stream writes, and exit behavior
 
-#### Scenario: Transaction domain drain remains
+#### Scenario: Protected-zone drain remains
 
 - **WHEN** a write-capable transform is implemented after this slice
-- **THEN** transformation transaction and protected-zone authority domains own
-  dry-run evidence, admitted write set, protected-zone checks, rollback data,
-  cleanup, and typed refusal records
+- **THEN** the transformation transaction domain owns dry-run evidence,
+  transaction input resolution, and typed refusal records
+- **AND** protected-zone authority, rollback data, cleanup, and write resources
+  move only through later named drains
 - **AND** provider command success alone is not sufficient to authorize writes
