+++ openspec/changes/deep-habitat-effect-transformation-transaction-domain/workstream/phase-record.md
@@ -3,31 +3,29 @@
 ## Phase
 
 - Project: Habitat Harness deep refactor
-- Phase: Fix service module
+- Phase: Transformation transaction domain drain
 - Owner: Effect-first implementation lane
-- Branch/Graphite stack: `agent-DRA-effect-fix-service-module` stacked on `agent-DRA-effect-hook-service-module`
+- Branch/Graphite stack: `agent-DRA-effect-transformation-transaction-domain` stacked on `agent-DRA-effect-verify-graph-cutover`
 - Started: 2026-06-19
 - Last updated: 2026-06-20
-- Status: implementation complete; review repairs and Graphite submit pending
+- Status: implementation complete; Graphite submit pending
 
 ## Objective
 
-- Target movement: make `habitat fix` an owned Effect-oRPC service module before
-  the deeper transformation transaction and protected-zone domain drain.
-- Non-goals: new live-write behavior.
-- Done condition: fix CLI routes through the service client, fix service
-  owns command-level orchestration, the obsolete `lib/fix` wrapper is removed,
-  current stream/exit behavior is preserved, and tests pin intent projection
-  plus architecture boundaries.
+- Target movement: make transformation transaction contracts, refusals,
+  transaction input resolution, rendering, and worktree observation a named
+  domain under `src/domains/transformation-transaction/**`.
+- Non-goals: new live-write behavior or protected-zone authority movement.
+- Done condition: `src/lib/pattern-apply/**` is deleted with no wrapper or
+  facade, fix and transactions service modules consume the domain directly, and
+  current dry-run/refusal behavior stays stable.
 
 ## Verification
 
 - `bun run --cwd tools/habitat-harness check` - passed.
-- `bun run --cwd tools/habitat-harness test -- test/service/fix-service.test.ts test/service/service-architecture.test.ts test/commands/habitat-commands.test.ts test/lib/pattern-apply.test.ts` - passed.
+- `bun run --cwd tools/habitat-harness test -- test/lib/pattern-apply.test.ts test/service/transactions-service.test.ts test/service/fix-service.test.ts test/service/service-architecture.test.ts` - passed.
 - `bun run --cwd tools/habitat-harness test` - passed.
 - `bun run biome:ci` - passed.
 - `bun run openspec -- validate deep-habitat-effect-transformation-transaction-domain --strict` - passed.
 - `bun run openspec:validate` - passed.
 - `git diff --check` - passed.
-- Evidence boundary: service-module slice; transformation transaction and
-  protected-zone authority drainage remains a later implementation unit.
