+++ tools/habitat-harness/src/domains/transformation-transaction/worktree.ts
@@ -1,4 +1,4 @@
-import { readGitState } from "../git-state.js";
+import { readGitState } from "../../lib/git-state.js";
 import { parseWorktreeObservation, type WorktreeObservation } from "./schema.js";
 
 export function observeWorktree(cwd?: string): WorktreeObservation {
