+++ tools/habitat-harness/src/domains/workspace-graph-integration/index.ts
@@ -35,6 +35,7 @@ export type {
   WorkspacePathClassification,
 } from "./classify.js";
 export {
+  ClassifyResultSchema,
   classifyPath,
   classifyPathResult,
   classifyTarget,
