+++ tools/habitat-harness/src/service/modules/fix/context.ts
@@ -3,7 +3,7 @@ import type {
   ApplyAdmission,
   ApplyTransactionInput,
 } from "../../../domains/pattern-governance/index.js";
-import type { WorktreeObservation } from "../../../lib/pattern-apply/index.js";
+import type { WorktreeObservation } from "../../../domains/transformation-transaction/index.js";
 import type { GritProvider } from "../../../providers/grit/index.js";
 
 export interface FixServiceOptions {
