+++ tools/habitat-harness/src/service/modules/transactions/contract.ts
@@ -3,7 +3,7 @@ import { eoc } from "effect-orpc";
 import {
   PatternApplyRecordSchema,
   PatternApplyRequestSchema,
-} from "../../../lib/pattern-apply/schema.js";
+} from "../../../domains/transformation-transaction/schema.js";
 import { type HabitatServiceErrorMap, habitatServiceErrorMap } from "../../errors.js";
 import type { HabitatServiceProcedureMeta } from "../../metadata.js";
 import { toStandardSchema } from "../../typebox-standard-schema.js";
